mypy: fix remaining errors in discovery/_fit.py, discovery/__init__.py, transforms domain_check typing, ensemble/_stackers.py, monitoring.py, missing.py, and misc float/str/bool no-any-return sites (135 -> 112 in-scope)
Types the self parameter on standalone functions bound onto CompositeTargetDiscovery (fit, _auto_base) so mypy can verify their "return self" / return-value contracts, adds the missing class-level attribute declarations those functions assign, removes a duplicate _auto_base_pool declaration, and wraps a handful of numeric/str/bool returns that were losing their concrete type through numpy/pandas call chains.

+++ src/mlframe/training/composite/discovery/_eval_stats.py
@@ -181,7 +181,7 @@ def bootstrap_gain_p_value(boot_gains: np.ndarray) -> float:
     if n == 0:
         return 1.0
     n_le0 = int(np.count_nonzero(g <= 0.0))
-    return (n_le0 + 1.0) / (n + 1.0)
+    return float((n_le0 + 1.0) / (n + 1.0))
 
 
 def benjamini_hochberg_reject(p_values, alpha: float) -> np.ndarray:
